Fixed issue with command interpreter queuing inputs from previous round.

+++ Assets/Scripts/System/GameState.cs
@@ -7,7 +7,7 @@ public class GameState
 
     public short RoundCount;
 
-    public ushort FrameCount;
+    public uint FrameCount;
 
     public ushort RoundTime;
